+++ source/controllers/text/TextControllerSub.cpp
@@ -267,7 +267,7 @@ void TextControllerSub::drawText(const std::string& text, int startX, int startY
 
 void TextControllerSub::clearScreen()
 {
-    dmaFillHalfWords(0, videoBuffer, 256 * 256 * sizeof(uint16_t)); // Clear the video buffer
+    dmaFillHalfWords(0, videoBuffer, 256 * 256 * sizeof(uint8_t)); // Clear the video buffer
 }
 
 // Helper Functions ======================================================
